// app.client/src/api/models/IncomeTypeListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IncomeTypeListVM = {
    incomeTypeID?: string | null;
    name?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
    incomeCategoryId?: string | null;
    incomeCategoryName?: string | null;
};